# openapi.yaml
openapi: 3.0.3
info:
  title: API Pedidos e Gestão
  version: 1.0.0
  description: >
    API para gerenciamento de usuários, produtos, unidades e pedidos.
    Documentação gerada a partir dos arquivos de rota fornecidos (usuarios.js, produtos.js, unidades.js, pedidos.js).
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UsuarioCreate:
      type: object
      required:
        - nome
        - senha
        - consentimentoLGPD
      properties:
        nome:
          type: string
          example: "Ana Silva"
        email:
          type: string
          format: email
          example: "ana@example.com"
        telefone:
          type: string
          example: "33988332244"
        senha:
          type: string
          format: password
          example: "senha123"
        consentimentoLGPD:
          type: boolean
          example: true
    UsuarioLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        telefone:
          type: string
        senha:
          type: string
          format: password
      example:
        email: "admin@example.com"
        senha: "1234"
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: "Ana Silva"
        email:
          type: string
          format: email
          example: "ana@example.com"
        telefone:
          type: string
          example: "33988332244"
        role:
          type: string
          enum: [CLIENTE, ADMIN, ATENDENTE, COZINHA, ENTREGADOR]
          example: CLIENTE
...
  responses:
    Unauthorized:
      description: Token inválido ou ausente
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Token inválido"
    Forbidden:
      description: Acesso negado (role insuficiente)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Sem permissão para acessar este recurso"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Recurso não encontrado"
    ServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Erro ao conectar: ..."
paths:
  /usuarios/register:
    post:
      tags: [Usuarios]
      summary: Registrar novo usuário (consentimento LGPD obrigatório)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioCreate'
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuário registrado com sucesso"
        '400':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /usuarios/login:
    post:
      tags: [Usuarios]
      summary: Autenticar usuário (email ou telefone)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioLogin'
      responses:
        '200':
          description: Token JWT e dados do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login realizado com sucesso"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '401':
          description: Senha inválida
        '403':
          description: Conta bloqueada por muitas tentativas
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar todos os usuários (ADMIN, ATENDENTE)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuarios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Usuarios]
      summary: Buscar usuário por ID (ADMIN, ATENDENTE)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Usuarios]
      summary: Atualizar usuário (ADMIN ou o próprio usuário)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: { type: string }
                email: { type: string }
                telefone: { type: string }
      responses:
        '200':
          description: Usuário atualizado com sucesso
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Usuarios]
      summary: Excluir conta (ADMIN pode excluir qualquer, CLIENTE pode excluir a própria)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conta excluída com sucesso
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /usuarios/register-admin:
    post:
      tags: [Usuarios]
      summary: Registrar funcionário (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, senha, consentimentoLGPD]
              properties:
                nome: { type: string }
                email: { type: string }
                telefone: { type: string }
                senha: { type: string }
                role:
                  type: string
                  enum: [ADMIN, ATENDENTE, COZINHA, ENTREGADOR]
                consentimentoLGPD: { type: boolean }
      responses:
        '201':
          description: Funcionário registrado
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuarios/{id}/fidelidade:
    get:
      tags: [Usuarios]
      summary: Consultar pontos de fidelidade de um cliente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pontos de fidelidade
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuario:
                    $ref: '#/components/schemas/Usuario'
                  pontos:
                    type: integer
                    example: 0
        '404':
          $ref: '#/components/responses/NotFound'
...